+++ lighthouse/classes/services/mongo.py
@@ -3,7 +3,7 @@ class ServiceMongoMixin(object):
         ...
 
     def get_samples_from_mongo(self, samples):
-        return []
+        ...
 
     def get_source_plate_uuid(self, barcode):
-        return "1234"
+        ...
